dogtor_api/requirements.txt

Django==4.2.5
djangorestframework==3.14.0
psycopg==3.1.12
python-dotenv==1.0.0

